Split Rust map and render modules
